@@ -9,29 +9,25 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "axios": "^1.7.7",
+    "axios": "^1.7.9",
     "formik": "^2.4.6",
     "i18next": "^24.2.2",
-    "i18next-browser-languagedetector": "^8.0.2",
-    "i18next-http-backend": "^3.0.2",
     "jwt-decode": "^4.0.0",
-    "next": "15.0.3",
-    "next-i18n-router": "^5.5.1",
+    "next": "15.1.6",
     "next-i18next": "^15.4.1",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-i18next": "^15.4.0",
-    "react-icons": "^5.3.0",
     "socket.io-client": "^4.8.1",
     "yup": "^1.6.1",
-    "zustand": "^5.0.1"
+    "zustand": "^5.0.3"
   },
   "devDependencies": {
-    "@types/node": "^22",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/node": "^20",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
     "postcss": "^8",
-    "tailwindcss": "^3.4.14",
+    "tailwindcss": "^3.4.1",
     "typescript": "^5"
   }
 }
