@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from "react";
+import { JSX, useEffect, useRef, useState } from "react";
 
 export enum ModalSizeEnum {
   XLARGE = "XLARGE",
